+++ packages/voice/src/models/v2/recording-format-type/recording-format-type.ts
@@ -1,0 +1,4 @@
+/**
+ * Audio format for this recording.
+ */
+export type RecordingFormatType = 'MP3' | 'WAV' | string;