+++ packages/voice/src/models/v2/svaml-commands/svaml-command/index.ts
@@ -1,0 +1,1 @@
+export type { SvamlCommand } from './svaml-command';